feat(): updating linking schema
