+++ README.md
@@ -8,3 +8,5 @@ https://docs.expo.dev/push-notifications/push-notifications-setup/
 If the project is configured EAS project secret GOOGLE_SERVICES_JSON has to be included uploading google-services.json downloaded from Firebase project.
 
 After installing apk or ipa in devices and start expo server dev client... an Expo Push Token is given like that: `ExponentPushToken[XXXXXXXXXXXXXX]`. It could be used for send push notifications to dev.
+
+Deep links has an schema like : `easbuildtest://`
