+++ android/app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         <action android:name="android.intent.action.VIEW"/>
         <category android:name="android.intent.category.DEFAULT"/>
         <category android:name="android.intent.category.BROWSABLE"/>
-        <data android:scheme="myapp"/>
+        <data android:scheme="easbuildtest"/>
         <data android:scheme="com.elenitaex5.easbuildtest"/>
       </intent-filter>
     </activity>
